Split completions in multiple files, move them to src/completions
