+++ src/completions.ts
@@ -6,22 +6,23 @@ On each input event, call updateCompletions on the array. That will mutate the a
 
 How to handle cached e.g. buffer information going out of date?
 
+Concrete completion classes have been moved to src/completions/.
+
 */
 
 import * as Fuse from "fuse.js"
 import { enumerate } from "./itertools"
 import { toNumber } from "./convert"
-import * as Messaging from "./messaging"
-import * as config from "./config"
-import { browserBg } from "./lib/webext"
 
-const DEFAULT_FAVICON = browser.extension.getURL("static/defaultFavicon.svg")
+export const DEFAULT_FAVICON = browser.extension.getURL(
+    "static/defaultFavicon.svg",
+)
 
 // {{{ INTERFACES
 
 type OptionState = "focused" | "hidden" | "normal"
 
-abstract class CompletionOption {
+export abstract class CompletionOption {
     /** What to fill into cmdline */
     value: string
     /** Control presentation of the option */
@@ -65,7 +66,7 @@ export abstract class CompletionSource {
 
 // Default classes
 
-abstract class CompletionOptionHTML extends CompletionOption {
+export abstract class CompletionOptionHTML extends CompletionOption {
     public html: HTMLElement
     public value
 
@@ -97,18 +98,18 @@ abstract class CompletionOptionHTML extends CompletionOption {
     }
 }
 
-interface CompletionOptionFuse extends CompletionOptionHTML {
+export interface CompletionOptionFuse extends CompletionOptionHTML {
     // For fuzzy matching
     fuseKeys: any[]
 }
 
-type ScoredOption = {
+export type ScoredOption = {
     index: number
     option: CompletionOptionFuse
     score: number
 }
 
-abstract class CompletionSourceFuse extends CompletionSource {
+export abstract class CompletionSourceFuse extends CompletionSource {
     public node
     public options: CompletionOptionFuse[]
     protected lastExstr: string
@@ -305,436 +306,6 @@ abstract class CompletionSourceFuse extends CompletionSource {
 }
 
 // }}}
-
-// {{{ IMPLEMENTATIONS
-
-class HistoryCompletionOption extends CompletionOptionHTML
-    implements CompletionOptionFuse {
-    public fuseKeys = []
-
-    constructor(public value: string, page: browser.history.HistoryItem) {
-        super()
-        if (!page.title) {
-            page.title = new URL(page.url).host
-        }
-
-        // Push properties we want to fuzmatch on
-        this.fuseKeys.push(page.title, page.url) // weight by page.visitCount
-
-        // Create HTMLElement
-        // need to download favicon
-        const favIconUrl = DEFAULT_FAVICON
-        // const favIconUrl = tab.favIconUrl ? tab.favIconUrl : DEFAULT_FAVICON
-        this.html = html`<tr class="HistoryCompletionOption option">
-            <td class="prefix">${"".padEnd(2)}</td>
-            <td></td>
-            <td>${page.title}</td>
-            <td><a class="url" target="_blank" href=${page.url}>${
-            page.url
-        }</a></td>
-        </tr>`
-    }
-}
-
-class BmarkCompletionOption extends CompletionOptionHTML
-    implements CompletionOptionFuse {
-    public fuseKeys = []
-
-    constructor(
-        public value: string,
-        bmark: browser.bookmarks.BookmarkTreeNode,
-    ) {
-        super()
-        if (!bmark.title) {
-            bmark.title = new URL(bmark.url).host
-        }
-
-        // Push properties we want to fuzmatch on
-        this.fuseKeys.push(bmark.title, bmark.url)
-
-        // Create HTMLElement
-        // need to download favicon
-        const favIconUrl = DEFAULT_FAVICON
-        // const favIconUrl = tab.favIconUrl ? tab.favIconUrl : DEFAULT_FAVICON
-        this.html = html`<tr class="HistoryCompletionOption option">
-            <td class="prefix">${"".padEnd(2)}</td>
-            <td></td>
-            <td>${bmark.title}</td>
-            <td><a class="url" target="_blank" href=${bmark.url}>${
-            bmark.url
-        }</a></td>
-        </tr>`
-    }
-}
-
-function sleep(ms: number) {
-    return new Promise(resolve => setTimeout(resolve, ms))
-}
-
-export class BmarkCompletionSource extends CompletionSourceFuse {
-    public options: BmarkCompletionOption[]
-
-    constructor(private _parent) {
-        super(["bmarks "], "BmarkCompletionSource", "Bookmarks")
-
-        this._parent.appendChild(this.node)
-    }
-
-    public async filter(exstr: string) {
-        this.lastExstr = exstr
-        const [prefix, query] = this.splitOnPrefix(exstr)
-
-        // Hide self and stop if prefixes don't match
-        if (prefix) {
-            // Show self if prefix and currently hidden
-            if (this.state === "hidden") {
-                this.state = "normal"
-            }
-        } else {
-            this.state = "hidden"
-            return
-        }
-
-        this.options = (await this.scoreOptions(query, 10)).map(
-            page => new BmarkCompletionOption(page.url, page),
-        )
-
-        this.updateChain()
-    }
-
-    updateChain() {
-        // Options are pre-trimmed to the right length.
-        this.options.forEach(option => (option.state = "normal"))
-
-        // Call concrete class
-        this.updateDisplay()
-    }
-
-    onInput() {}
-
-    private async scoreOptions(query: string, n: number) {
-        // Search bookmarks, dedupe and sort by frecency
-        let bookmarks = await browserBg.bookmarks.search({ query })
-        bookmarks = bookmarks.filter(b => {
-            try {
-                return new URL(b.url)
-            } catch (e) {
-                return false
-            }
-        })
-
-        bookmarks.sort((a, b) => b.dateAdded - a.dateAdded)
-
-        return bookmarks.slice(0, n)
-    }
-
-    select(option: CompletionOption) {
-        if (this.lastExstr !== undefined && option !== undefined) {
-            this.completion = "open " + option.value
-            option.state = "focused"
-            this.lastFocused = option
-        } else {
-            throw new Error("lastExstr and option must be defined!")
-        }
-    }
-}
-
-export class HistoryCompletionSource extends CompletionSourceFuse {
-    public options: HistoryCompletionOption[]
-
-    constructor(private _parent) {
-        super(
-            ["open ", "tabopen ", "winopen "],
-            "HistoryCompletionSource",
-            "History",
-        )
-
-        this._parent.appendChild(this.node)
-    }
-
-    public async filter(exstr: string) {
-        this.lastExstr = exstr
-        const [prefix, query] = this.splitOnPrefix(exstr)
-
-        // Hide self and stop if prefixes don't match
-        if (prefix) {
-            // Show self if prefix and currently hidden
-            if (this.state === "hidden") {
-                this.state = "normal"
-            }
-        } else {
-            this.state = "hidden"
-            return
-        }
-
-        this.options = (await this.scoreOptions(query, 10)).map(
-            page => new HistoryCompletionOption(page.url, page),
-        )
-
-        this.updateChain()
-    }
-
-    updateChain() {
-        // Options are pre-trimmed to the right length.
-        this.options.forEach(option => (option.state = "normal"))
-
-        // Call concrete class
-        this.updateDisplay()
-    }
-
-    onInput() {}
-
-    private frecency(item: browser.history.HistoryItem) {
-        // Doesn't actually care about recency yet.
-        return item.visitCount * -1
-    }
-
-    private async scoreOptions(query: string, n: number) {
-        const newtab = browser.runtime.getManifest()["chrome_url_overrides"]
-            .newtab
-        const newtaburl = browser.extension.getURL(newtab)
-        if (!query) {
-            return (await browserBg.topSites.get())
-                .filter(page => page.url !== newtaburl)
-                .slice(0, n)
-        } else {
-            // Search history, dedupe and sort by frecency
-            let history = await browserBg.history.search({
-                text: query,
-                maxResults: Number(config.get("historyresults")),
-                startTime: 0,
-            })
-
-            // Remove entries with duplicate URLs
-            const dedupe = new Map()
-            for (const page of history) {
-                if (page.url !== newtaburl) {
-                    if (dedupe.has(page.url)) {
-                        if (
-                            dedupe.get(page.url).title.length <
-                            page.title.length
-                        ) {
-                            dedupe.set(page.url, page)
-                        }
-                    } else {
-                        dedupe.set(page.url, page)
-                    }
-                }
-            }
-            history = [...dedupe.values()]
-
-            history.sort((a, b) => this.frecency(a) - this.frecency(b))
-
-            return history.slice(0, n)
-        }
-    }
-}
-
-class BufferCompletionOption extends CompletionOptionHTML
-    implements CompletionOptionFuse {
-    public fuseKeys = []
-
-    constructor(
-        public value: string,
-        tab: browser.tabs.Tab,
-        public isAlternative = false,
-    ) {
-        super()
-        // Two character buffer properties prefix
-        let pre = ""
-        if (tab.active) pre += "%"
-        else if (isAlternative) {
-            pre += "#"
-            this.value = "#"
-        }
-        if (tab.pinned) pre += "@"
-
-        // Push prefix before padding so we don't match on whitespace
-        this.fuseKeys.push(pre)
-
-        // Push properties we want to fuzmatch on
-        this.fuseKeys.push(String(tab.index + 1), tab.title, tab.url)
-
-        // Create HTMLElement
-        const favIconUrl = tab.favIconUrl ? tab.favIconUrl : DEFAULT_FAVICON
-        this.html = html`<tr class="BufferCompletionOption option">
-            <td class="prefix">${pre.padEnd(2)}</td>
-            <td><img src=${favIconUrl} /></td>
-            <td>${tab.index + 1}: ${tab.title}</td>
-            <td><a class="url" target="_blank" href=${tab.url}>${
-            tab.url
-        }</a></td>
-        </tr>`
-    }
-}
-
-export class BufferCompletionSource extends CompletionSourceFuse {
-    public options: BufferCompletionOption[]
-
-    // TODO:
-    //     - store the exstr and trigger redraws on user or data input without
-    //       callback faffery
-    //     - sort out the element redrawing.
-
-    constructor(private _parent) {
-        super(
-            ["buffer ", "tabclose ", "tabdetach ", "tabduplicate ", "tabmove "],
-            "BufferCompletionSource",
-            "Buffers",
-        )
-
-        this.updateOptions()
-        this._parent.appendChild(this.node)
-    }
-
-    private async updateOptions(exstr?: string) {
-        /* console.log('updateOptions', this.optionContainer) */
-        const tabs: browser.tabs.Tab[] = await Messaging.message(
-            "commandline_background",
-            "currentWindowTabs",
-        )
-
-        const options = []
-
-        // Get alternative tab, defined as last accessed tab.
-        const alt = tabs.sort((a, b) => {
-            return a.lastAccessed < b.lastAccessed ? 1 : -1
-        })[1]
-        tabs.sort((a, b) => {
-            return a.index < b.index ? -1 : 1
-        })
-
-        for (const tab of tabs) {
-            options.push(
-                new BufferCompletionOption(
-                    (tab.index + 1).toString(),
-                    tab,
-                    tab === alt,
-                ),
-            )
-        }
-
-        /* console.log('updateOptions end', this.waiting, this.optionContainer) */
-        this.options = options
-        this.updateChain()
-    }
-
-    async onInput(exstr) {
-        // Schedule an update, if you like. Not very useful for buffers, but
-        // will be for other things.
-        this.updateOptions()
-    }
-
-    setStateFromScore(scoredOpts: ScoredOption[]) {
-        super.setStateFromScore(scoredOpts, true)
-    }
-
-    /** Score with fuse unless query is a single # or looks like a buffer index */
-    scoredOptions(query: string, options = this.options): ScoredOption[] {
-        const args = query.trim().split(/\s+/gu)
-        if (args.length === 1) {
-            // if query is an integer n and |n| < options.length
-            if (Number.isInteger(Number(args[0]))) {
-                let index = Number(args[0]) - 1
-                if (Math.abs(index) < options.length) {
-                    index = index.mod(options.length)
-                    return [
-                        {
-                            index,
-                            option: options[index],
-                            score: 0,
-                        },
-                    ]
-                }
-            } else if (args[0] === "#") {
-                for (const [index, option] of enumerate(options)) {
-                    if (option.isAlternative) {
-                        return [
-                            {
-                                index,
-                                option,
-                                score: 0,
-                            },
-                        ]
-                    }
-                }
-            }
-        }
-
-        // If not yet returned...
-        return super.scoredOptions(query, options)
-    }
-}
-
-class BufferAllCompletionOption extends CompletionOptionHTML {
-    public fuseKeys = []
-
-    constructor(public value: string, tab: browser.tabs.Tab, winindex: number) {
-        super()
-        this.fuseKeys.push(String(tab.id), tab.title, tab.url)
-
-        // Create HTMLElement
-        const favIconUrl = tab.favIconUrl ? tab.favIconUrl : DEFAULT_FAVICON
-        this.html = html`<tr class="BufferAllCompletionOption option">
-            <td class="prefix"></td>
-            <td><img src=${favIconUrl} /></td>
-            <td>${winindex}.${tab.index + 1}: ${tab.title}</td>
-            <td><a class="url" target="_blank" href=${tab.url}>${
-            tab.url
-        }</a></td>
-        </tr>`
-    }
-}
-
-export class BufferAllCompletionSource extends CompletionSourceFuse {
-    public options: BufferAllCompletionOption[]
-
-    constructor(private _parent) {
-        super(["bufferall "], "BufferAllCompletionSource", "All Buffers")
-
-        this.updateOptions()
-        this._parent.appendChild(this.node)
-    }
-
-    async onInput(exstr) {
-        this.updateOptions()
-    }
-
-    private async updateOptions(exstr?: string) {
-        const tabs: browser.tabs.Tab[] = await Messaging.message(
-            "commandline_background",
-            "allWindowTabs",
-        )
-
-        const options = []
-
-        tabs.sort((a, b) => {
-            if (a.windowId == b.windowId) return a.index - b.index
-            return a.windowId - b.windowId
-        })
-
-        // Window Ids don't make sense so we're using LASTID and IDCOUNT to compute a window index
-        // This relies on the fact that tabs are sorted by window ids
-        let lastId = 0
-        let index = 0
-        for (const tab of tabs) {
-            if (lastId != tab.windowId) {
-                lastId = tab.windowId
-                index += 1
-            }
-            options.push(
-                new BufferAllCompletionOption(tab.id.toString(), tab, index),
-            )
-        }
-
-        this.options = options
-        this.updateChain()
-    }
-
-    setStateFromScore(scoredOpts: ScoredOption[]) {
-        super.setStateFromScore(scoredOpts, true)
-    }
-}
 
 // {{{ UNUSED: MANAGING ASYNC CHANGES
